Create superclass `Contact` and implements inheritance
